refactor authenticated actions via Policy

+++ app/Http/Controllers/QuestionController.php
@@ -53,7 +53,7 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
-        $this->ensureUserOwnsQuestion($question);
+        $this->authorize('update', $question);
 
         return view('questions.edit', compact('question'));
     }
@@ -63,7 +63,7 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
-        $this->ensureUserOwnsQuestion($question);
+        $this->authorize('update', $question);
 
         $question->update($request->only('title', 'body'));
 
@@ -75,15 +75,10 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
-        $this->ensureUserOwnsQuestion($question);
+        $this->authorize('delete', $question);
 
         $question->delete();
 
         return redirect('/questions')->with('success', "Your question has been deleted.");
-    }
-
-    private function ensureUserOwnsQuestion(Question $question): void
-    {
-        abort_unless($question->user_id === auth()->id(), 403);
     }
 }
